requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.7.0
python-dotenv>=1.0.0
lxml>=4.9.0
feedparser>=6.0.0
